split rest-api.js: split by api (except study api)
