+++ .env
@@ -1,3 +1,4 @@
 REACT_APP_API_GATEWAY=api/gateway
 REACT_APP_WS_GATEWAY=ws/gateway
+REACT_APP_API_VERSION=v1
 EXTEND_ESLINT=true
