+++ src/components/dialogs/line-types-catalog/line-type-segment-form.js
@@ -27,7 +27,6 @@ import {
 import LineTypesCatalogSelectorDialog from './line-types-catalog-selector-dialog';
 import { roundToDefaultPrecision } from '../../../utils/rounding';
 import LineTypeSegmentCreation from './line-type-segment-creation';
-import { getLineTypesCatalog } from '../../../utils/rest-api';
 import { emptyLineSegment } from './line-type-segment-dialog';
 import {
     calculateResistance,
@@ -36,6 +35,7 @@ import {
 } from '../../utils/utils';
 import makeStyles from '@mui/styles/makeStyles';
 import { useSnackMessage } from '@gridsuite/commons-ui';
+import { getLineTypesCatalog } from '../../../services/network-modification';
 
 const useStyles = makeStyles((theme) => ({
     header: {
