+++ src/components/graph/menus/network-modification-node-editor.js
@@ -10,7 +10,6 @@ import {
     changeNetworkModificationOrder,
     copyOrMoveModifications,
     deleteModifications,
-    fetchNetworkModification,
     fetchNetworkModifications,
 } from '../../../utils/rest-api';
 import { useSnackMessage } from '@gridsuite/commons-ui';
@@ -65,6 +64,7 @@ import { UPDATE_TYPE } from 'components/network/constants';
 import { FetchStatus } from 'utils/rest-api';
 import LineSplitWithVoltageLevelDialog from 'components/dialogs/network-modifications/line-split-with-voltage-level/line-split-with-voltage-level-dialog';
 import TwoWindingsTransformerModificationDialog from '../../dialogs/network-modifications/two-windings-transformer/modification/two-windings-transformer-modification-dialog';
+import { fetchNetworkModification } from '../../../services/network-modification';
 
 const useStyles = makeStyles((theme) => ({
     listContainer: {
