+++ src/services/config.js
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+import { API_VERSION, backendFetch, backendFetchJson } from '../utils/rest-api';
+import { getAppName } from '../utils/config-params';
+
+const PREFIX_CONFIG_QUERIES = `${process.env.REACT_APP_API_GATEWAY}/config`;
+
+export function fetchConfigParameters(appName) {
+    console.info('Fetching UI configuration params for app : ' + appName);
+    const fetchParams =
+        PREFIX_CONFIG_QUERIES +
+        `/${API_VERSION}/applications/${appName}/parameters`;
+    return backendFetchJson(fetchParams);
+}
+
+export function fetchConfigParameter(name) {
+    const appName = getAppName(name);
+    console.info(
+        "Fetching UI config parameter '%s' for app '%s' ",
+        name,
+        appName
+    );
+    const fetchParams =
+        PREFIX_CONFIG_QUERIES +
+        `/${API_VERSION}/applications/${appName}/parameters/${name}`;
+    return backendFetch(fetchParams).then((response) =>
+        response.status === 204 ? null : response.json()
+    );
+}
+
+export function updateConfigParameter(name, value) {
+    const appName = getAppName(name);
+    console.info(
+        "Updating config parameter '%s=%s' for app '%s' ",
+        name,
+        value,
+        appName
+    );
+    const updateParams =
+        PREFIX_CONFIG_QUERIES +
+        `/v1/applications/${appName}/parameters/${name}?value=` +
+        encodeURIComponent(value);
+    return backendFetch(updateParams, { method: 'put' });
+}
